requirements.txt
requests
tenacity
python-dotenv
pandas
numpy
psycopg2-binary
sqlalchemy
structlog
click
tqdm